Rename WTM webextension-packages to reporting
addresses https://github.com/ghostery/ghostery-extension/issues/1723
