+++ extension-manifest-v2/package.json
@@ -46,7 +46,7 @@
     "@ghostery/trackers-preview": "^1.0.0",
     "@ghostery/ui": "^1.0.0",
     "@sentry/browser": "7.1.1",
-    "@whotracksme/webextension-packages": "^5.0.16",
+    "@whotracksme/reporting": "^5.1.17",
     "classnames": "^2.3.2",
     "d3": "^7.8.5",
     "foundation-sites": "^6.6.2",
