+++ extension-manifest-v3/package.json
@@ -41,7 +41,7 @@
     "@ghostery/trackers-preview": "^1.0.0",
     "@ghostery/ui": "^1.0.0",
     "@github/relative-time-element": "^4.3.0",
-    "@whotracksme/webextension-packages": "^5.0.16",
+    "@whotracksme/reporting": "^5.1.17",
     "hybrids": "^9.0.0",
     "idb": "^7.1.1",
     "jwt-decode": "^4.0.0",
