+++ extension-manifest-v3/src/content_scripts/whotracksme/reporting.js
@@ -1,4 +1,4 @@
-import { recordMouseDown } from '@whotracksme/webextension-packages/packages/reporting/src/request/content-script.js';
+import { recordMouseDown } from '@whotracksme/reporting/reporting/src/request/content-script.js';
 
 window.addEventListener('mousedown', (ev) => {
   const { event, context, href } = recordMouseDown(ev);
